refactor file downloading
