+++ src/services/schemaService.utils.ts
@@ -1,4 +1,4 @@
-import { getConvertedFile, getSchemaById } from './schemaApiService';
+import { getSchemaById } from './schemaApiService';
 
 export const downloadTurtleFile = async (id: string) => {
   const response = await getSchemaById(id);
@@ -10,25 +10,4 @@ export const downloadTurtleFile = async (id: string) => {
   document.body.appendChild(element);
   element.click();
   document.body.removeChild(element);
-}
-
-export const downloadJsonFile = async (id: string) => {
-  try {
-    const response = await getConvertedFile(id);
-    const filename = id + '.json';
-
-    const blob = new Blob([JSON.stringify(response, null, 2)], { type: 'application/json' });
-
-    const element = document.createElement('a');
-    element.href = URL.createObjectURL(blob);
-    element.download = filename;
-
-    document.body.appendChild(element);
-
-    element.click();
-
-    document.body.removeChild(element);
-  } catch (error) {
-    console.error('Error:', error);
-  }
 };
